Deprecate
The tool does not work with newer Waf versions. Investigating the new
Waf code base, a re-implementation is found to be not quite worthwhile.

It would involve ripping out a large part of the involved methods in the
BuildContext and related classes just to replace them with variants that
are apart from a few words identical to the originals. This would leave
the tool very unstable to changes to the patched code. As this is a
central part of Waf, often touched by many workflows, this is likely.

So, a new installdir tool would remain unstable and difficult to adapt,
making it a dangerous filed for both author and user. It is thus more
advisable for users to switch to alternatives described to the README in
this commit.